Edit transport rules, changed to draw
