+++ src/pages/email/transport/list-rules/index.js
@@ -1,15 +1,24 @@
 import { Layout as DashboardLayout } from "/src/layouts/index.js";
 import { CippTablePage } from "/src/components/CippComponents/CippTablePage.jsx";
-import { Book, DoDisturb, Done } from "@mui/icons-material";
+import { Book, DoDisturb, Done, Edit } from "@mui/icons-material";
 import { TrashIcon } from "@heroicons/react/24/outline";
 import { CippAddTransportRuleDrawer } from "../../../../components/CippComponents/CippAddTransportRuleDrawer";
-import { Button } from "@mui/material";
-import { RocketLaunch } from "@mui/icons-material";
-import Link from "next/link";
+import { CippTransportRuleDrawer } from "../../../../components/CippComponents/CippTransportRuleDrawer";
+import { useSettings } from "/src/hooks/use-settings";
+import { useRef } from "react";
 
 const Page = () => {
   const pageTitle = "Transport Rules";
   const cardButtonPermissions = ["Exchange.TransportRule.ReadWrite"];
+  const tableRef = useRef();
+  const currentTenant = useSettings().currentTenant;
+
+  const handleRuleSuccess = () => {
+    // Refresh the table after successful create/edit
+    if (tableRef.current) {
+      tableRef.current.refreshData();
+    }
+  };
 
   const actions = [
     {
@@ -23,22 +32,41 @@ const Page = () => {
     {
       label: "Enable Rule",
       type: "POST",
-      url: "/api/EditTransportRule",
+      url: "/api/AddEditTransportRule",
       data: {
-        State: "!Enable",
-        GUID: "Guid",
+        Enabled: "!Enabled",
+        Identity: "Guid",
+        Name: "Name",
       },
+      condition: (row) => row.State === "Disabled",
       confirmText: "Are you sure you want to enable this rule?",
       icon: <Done />,
     },
     {
+      label: "Edit Rule",
+      customComponent: (row, {drawerVisible, setDrawerVisible}) => (
+        <CippTransportRuleDrawer
+          isEditMode={true}
+          ruleId={row.Guid}
+          requiredPermissions={cardButtonPermissions}
+          onSuccess={handleRuleSuccess}
+          drawerVisible={drawerVisible}
+          setDrawerVisible={setDrawerVisible}
+        />
+      ),
+      icon: <Edit />,
+      multiPost: false,
+    },
+    {
       label: "Disable Rule",
       type: "POST",
-      url: "/api/EditTransportRule",
+      url: "/api/AddEditTransportRule",
       data: {
-        State: "!Disable",
-        GUID: "Guid",
+        Enabled: "!Disabled",
+        Identity: "Guid",
+        Name: "Name",
       },
+      condition: (row) => row.State === "Enabled",
       confirmText: "Are you sure you want to disable this rule?",
       icon: <DoDisturb />,
     },
@@ -93,9 +121,11 @@ const Page = () => {
 
   return (
     <CippTablePage
+      ref={tableRef}
       title={pageTitle}
       apiUrl="/api/ListTransportRules"
       apiDataKey="Results"
+      queryKey= {`Transport Rules - ${currentTenant}`}
       actions={actions}
       offCanvas={offCanvas}
       simpleColumns={simpleColumns}
@@ -103,13 +133,12 @@ const Page = () => {
       cardButton={
         <>
           <CippAddTransportRuleDrawer requiredPermissions={cardButtonPermissions} />
-          <Button
-            component={Link}
-            href="/email/transport/new-rules/add"
-            startIcon={<RocketLaunch />}
-          >
-            New Transport Rule
-          </Button>
+          <CippTransportRuleDrawer
+            buttonText="New Transport Rule"
+            isEditMode={false}
+            requiredPermissions={cardButtonPermissions}
+            onSuccess={handleRuleSuccess}
+          />
         </>
       }
     />
